added code for an analog battery check
i added some code for checking the battery level using an analog pin and a mosfet. i also wanted to add some code to make it check if its charging, but for some reason, it kept LITERALLY JUST READING A 0 AS A 1 because i didnt even set up a pin to check if its charging and instead just set charging to "false", but it kept executing the code anyways!