Adding preprocessor directives to support ESP32-S2FN4R2 dev boards
Created a new board definition for ESP32-S2 mini dev boards

Conditionally changed the default Serial0 pins to ones that are available on the mini dev board. Other boards are not affected.

Added a preprocessor directive to exclude softwareserial in GpsInterface.h, as it caused compilation issues with ESP32-S2FN4R2 boards.